office365/pyproject.toml
[project]
name = "office365"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "jupyter",
    "langgraph-cli[inmem]",
    "langchain-community>=0.3.31",
    "langchain-openai>=0.3.35",
    "langgraph>=0.6.10",
    "o365>=2.1.7",
    "fastapi>=0.111.0",
    "ngrok>=1.4.0",
    "loguru>=0.7.2",
    "uvicorn>=0.29.0",
    "msal>=1.34.0",
    "msgraph-sdk>=1.46.0",
    "langchain-classic>=1.0.0",
    "pymysql>=1.1.2",
    "langchain>=1.0.2",
]
